VRMShaders.TextureExporter, AssetTextureUtil
