+++ Assets/UniGLTF/Editor/UniGLTF.Editor.asmdef
@@ -4,7 +4,8 @@
         "UniGLTF",
         "MeshUtility",
         "MeshUtility.Editor",
-        "VRMShaders"
+        "VRMShaders",
+        "VRMShaders.Editor"
     ],
     "optionalUnityReferences": [],
     "includePlatforms": [
