+++ Assets/UniGLTF/Tests/UniGLTF.Tests.asmdef
@@ -3,7 +3,8 @@
     "references": [
         "UniGLTF",
         "UniGLTF.Editor",
-        "VRMShaders"
+        "VRMShaders",
+        "VRMShaders.Editor"
     ],
     "optionalUnityReferences": [
         "TestAssemblies"
