+++ Assets/UniGLTF/Tests/UniGLTF/UniGLTFTests.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using UniJSON;
 using UnityEngine;
-
+using VRMShaders;
 
 namespace UniGLTF
 {
