+++ Assets/VRM.Samples/Editor/Tests/VRM.Samples.Editor.Tests.asmdef
@@ -8,7 +8,7 @@
         "UniVRM.Editor",
         "VRM.Tests",
         "VRMShaders",
-        "UniGLTF.Editor"
+        "VRMShaders.Editor"
     ],
     "optionalUnityReferences": [
         "TestAssemblies"
