+++ Assets/VRM/Tests/VRMLoadTests.cs
@@ -4,6 +4,7 @@ using System.IO;
 using NUnit.Framework;
 using UniGLTF;
 using UnityEngine;
+using VRMShaders;
 
 namespace VRM
 {
