+++ Assets/VRMShaders/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 58204522f08a6fe4f846a5817a37f902
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
