+++ Assets/VRMShaders/Editor/VRMShaders.Editor.asmdef
@@ -1,0 +1,3 @@
+﻿{
+	"name": "VRMShaders.Editor"
+}
