+++ Assets/VRMShaders/Runtime/TextureExporter.cs.meta
